+++ src/Entities/PayPalAccount.php
@@ -91,11 +91,21 @@ final class PayPalAccount extends Entity
     }
 
     /**
+     * @deprecated The method is deprecated and will be removed in next version. Please use getApprovalUrl.
+     *
      * @return string
      */
     public function getApprovalLink()
     {
-        return $this->getLink('approval_url');
+        return $this->getApprovalUrl();
+    }
+
+    /**
+     * @return null|string
+     */
+    public function getApprovalUrl()
+    {
+        return $this->getLink('approvalUrl');
     }
 
     /**